+++ LayerKit.embeddedframework/LayerKit.framework/Versions/A/Headers/LYRConstants.h
@@ -49,28 +49,6 @@ extern NSString *const LYRObjectChangePropertyKey; // i.e. participants, metadat
 extern NSString *const LYRObjectChangeOldValueKey; // The value before synchronization
 extern NSString *const LYRObjectChangeNewValueKey; // The value after synchronization
 
-///------------------
-/// @name Option Keys
-///------------------
-
-/**
- @abstract The option key used in the conversation to specify the metadata content that is synchronized across participants.
- The value associated to this key must be passed in the `options` dictionary argument when creating the conversation.
- */
-extern NSString *const LYRConversationOptionsMetadataKey;
-
-/**
- @abstract The option key used in the message to specify the APNS alert message the server should include in the push delivered to the receiver of the message.
- The value associated to this key must be passed in the `options` dictionary argument when creating the message.
- */
-extern NSString *const LYRMessageOptionsPushNotificationAlertKey;
-
-/**
- @abstract The option key used in the message to specify the APNS sound name the server should include in the push delivered to the receiver of the message.
- The value associated to this key must be passed in the `options` dictionary argument when creating the message.
- */
-extern NSString *const LYRMessageOptionsPushNotificationSoundNameKey;
-
 ///-----------------------
 /// @name Typing indicator
 ///-----------------------